Split Pagination into its own component, add aria-label for accessibility
