+++ src/views/VideosList.vue
@@ -21,38 +21,7 @@
         </div>
       </div>
       <div class="page-navigation" v-if="lastPage !== 1">
-        <button
-          id="pagination-go-to-start" class="secondary-button"
-          v-on:click="setPagination('start')"
-        >
-          First Page
-        </button>
-        <form id="buttonWithText" v-on:submit.prevent="actOnButtonInput">
-          <div id="slider">
-            <span v-if="navIsButton">
-              <input
-                class="third-button button-disabled collapsed" type="button" id="toggle"
-                :value="this.currentPage" v-on:click="actOnButtonInput()"
-              />
-            </span>
-            <span v-else>
-              <input
-                class="third-button button-disabled" type="number" id="input" tabindex="-1"
-                placeholder="#" size=10 v-model="inputedUserPage"
-              />
-              <input
-                class="third-button" type="submit" id="submitSwitchButton"
-                tabindex="-1" value="Go"
-              />
-            </span>
-          </div>
-        </form>
-        <button
-          id="pagination-go-to-end" class="secondary-button"
-          v-on:click="setPagination('end')"
-        >
-          Last Page
-        </button>
+        <Pagination />
       </div>
     </div>
   </div>
@@ -64,10 +33,12 @@ import { debounce } from 'debounce';
 import { Component, Vue } from 'vue-property-decorator';
 import { State, Getter } from 'vuex-class';
 import VideoComponent from '../components/VideoComponent.vue';
+import Pagination from '../components/Pagination.vue';
 
 @Component({
   components: {
     VideoComponent,
+    Pagination,
   },
 })
 export default class VideosList extends Vue {
@@ -75,25 +46,11 @@ export default class VideosList extends Vue {
 
   @State('flashMessage') flashMessage!: string;
 
-  @State('currentPage') currentPage!: number;
-
   @State('vidPerPage') vidPerPage!: number;
 
   @Getter('currentPageVideosList') currentPageVideosList!: [];
 
-  @Getter('lastPage') lastPage!: number;
-
-  navIsButton: boolean;
-
-  inputedUserPage: string;
-
   // https://github.com/ktsn/vuex-class/issues/19
-
-  constructor() {
-    super();
-    this.navIsButton = true;
-    this.inputedUserPage = '';
-  }
 
   mounted() {
     window.addEventListener('resize', debounce(this.configureImageNumberPerPage, 300));
@@ -111,33 +68,6 @@ export default class VideosList extends Vue {
     } else if (width < 1100 && this.$store.state.vidPerPage !== 8) {
       this.$store.commit('setVideoPerPage', 8);
     }
-  }
-
-  setPagination(page: string): void {
-    if (page === 'end') {
-      this.$store.commit('setCurrentPage', this.lastPage);
-    } else if (page === 'start') {
-      this.$store.commit('setCurrentPage', 1);
-    }
-  }
-
-  actOnButtonInput(): void{
-    if (this.navIsButton) {
-      this.navIsButton = !this.navIsButton;
-    } else if (this.validateDataIsAnAppropriateNumber(this.inputedUserPage)) {
-      this.$store.commit('setCurrentPage', Number(this.inputedUserPage));
-      this.navIsButton = !this.navIsButton;
-    }
-  }
-
-  validateDataIsAnAppropriateNumber(userData: string): boolean {
-    const inputedUserDataAsNumber = Number(userData);
-    return (
-      Number.isInteger(inputedUserDataAsNumber)
-      && inputedUserDataAsNumber > 0
-      && inputedUserDataAsNumber <= this.lastPage
-      && this.currentPage !== inputedUserDataAsNumber
-    );
   }
 
   returnToLastPage() {
@@ -184,42 +114,8 @@ svg
   text-align: left
   margin-left: 3em
 
-%local-button
-  padding: 10px 22px
-  margin-right: .4rem
-  margin-left: .4rem
-
-.secondary-button
-  @extend %local-button
-
-.third-button
-  @extend %local-button
-
 .page-navigation
   margin-top: 1em
   margin-bottom: 1em
 
-
-.button-disabled
-  cursor: text
-
-#buttonWithText
-  display: inline-block
-
-#slider
-  transition: 0.5s
-
-#slider.expanded
-  transform: translateX(-50%)
-
-
-#toggle
-  flex: 0 0 50%
-
-#input
-  flex: auto
-  max-width: 5em
-
-#ok
-  flex: none
 </style>
